refact : 족보 부자 가게 api 쿼리 튜닝

+++ src/main/java/com/matdori/matdori/repositoy/Dto/JokboRichStore.java
@@ -9,9 +9,9 @@ public class JokboRichStore {
     private Long storeIndex;
     private String name;
     private String imgUrl;
-    private Integer jokboCnt;
+    private Long jokboCnt;
 
-    public JokboRichStore(Long storeIndex, String name, String imgUrl, int jokboCnt) {
+    public JokboRichStore(Long storeIndex, String name, String imgUrl, Long jokboCnt) {
         this.storeIndex = storeIndex;
         this.name = name;
         this.imgUrl = imgUrl;
